rename new_fetcher to dereference_object_id
