+++ crates/apub/src/activities/community/add_mod.rs
@@ -9,7 +9,7 @@ use crate::{
   },
   activity_queue::send_to_community_new,
   extensions::context::lemmy_context,
-  fetcher::new_fetcher::dereference,
+  fetcher::dereference_object_id::dereference,
   generate_moderators_url,
   ActorType,
 };
