+++ crates/apub/src/activities/community/remove_mod.rs
@@ -10,7 +10,7 @@ use crate::{
   },
   activity_queue::send_to_community_new,
   extensions::context::lemmy_context,
-  fetcher::new_fetcher::dereference,
+  fetcher::dereference_object_id::dereference,
   generate_moderators_url,
   ActorType,
 };
