+++ crates/apub/src/activities/following/undo.rs
@@ -7,7 +7,7 @@ use crate::{
   },
   activity_queue::send_activity_new,
   extensions::context::lemmy_context,
-  fetcher::new_fetcher::dereference,
+  fetcher::dereference_object_id::dereference,
   ActorType,
 };
 use activitystreams::{
