+++ crates/apub/src/activities/post/create_or_update.rs
@@ -10,7 +10,7 @@ use crate::{
   },
   activity_queue::send_to_community_new,
   extensions::context::lemmy_context,
-  fetcher::new_fetcher::dereference,
+  fetcher::dereference_object_id::dereference,
   objects::{post::Page, FromApub, ToApub},
   ActorType,
 };
